+++ src/mlpack/methods/ann/layer/base_layer.hpp
@@ -24,6 +24,7 @@
 #include <mlpack/methods/ann/activation_functions/mish_function.hpp>
 #include <mlpack/methods/ann/activation_functions/lisht_function.hpp>
 #include <mlpack/methods/ann/activation_functions/gelu_function.hpp>
+#include <mlpack/methods/ann/activation_functions/elliot_function.hpp>
 #include <mlpack/methods/ann/activation_functions/elish_function.hpp>
 
 namespace mlpack {
@@ -46,6 +47,7 @@ namespace ann /** Artificial Neural Network. */ {
  *  - LiSHTLayer
  *  - GELULayer
  *  - ELiSHLayer
+ *  - ElliotLayer
  *
  * @tparam ActivationFunction Activation function used for the embedding layer.
  * @tparam InputDataType Type of the input data (arma::colvec, arma::mat,
@@ -238,6 +240,17 @@ template <
     typename OutputDataType = arma::mat
 >
 using GELUFunctionLayer = BaseLayer<
+    ActivationFunction, InputDataType, OutputDataType>;
+
+/**
+ * Standard Elliot-Layer using the Elliot activation function.
+ */
+template <
+    class ActivationFunction = ElliotFunction,
+    typename InputDataType = arma::mat,
+    typename OutputDataType = arma::mat
+>
+using ElliotFunctionLayer = BaseLayer<
     ActivationFunction, InputDataType, OutputDataType>;
 
 /**